Fix test

+++ notary-integration-test/src/integration-test/kotlin/integration/eth/contracts/MasterTest.kt
@@ -681,6 +681,6 @@ class MasterTest {
         val n = master.peersCount().send()
 
         Assertions.assertTrue(result.isStatusOK)
-        Assertions.assertEquals(3, n)
+        Assertions.assertEquals(3, n.toInt())
     }
 }
